+++ src/components/index.ts
@@ -1,0 +1,7 @@
+import CityPaint from './CityPaint';
+import CountyPaint from './CountyPaint';
+
+export {
+  CityPaint,
+  CountyPaint,
+};